Break out the levelOne class to its own file. Create a folder with utility classes and fonts
