+++ utils/fonts.js
@@ -1,0 +1,1 @@
+export const FONT_STYLE = "retroGame";
